First Major Overhaul Update
Finally getting around to making this work and maybe fun.
Added, fixed, or implemented the following commands: Exit, attack, use,
equip, unequip, help, enter, eat, and search. Added the parseInv command
to streamline, although it is clear the whole system is kind of silly,
and in the long term should be remade as a more modular system for
inputs.
Updated most of the specific commands for each type of area, and added
or improved upon the flavor text greatly. Made items begin to have their
uses. Added several ways to die. Cleaned up some of the default methods
and their implementations in classes which inherit them.