Un-Random Update
Main change is that I removed the random map generation, and begun work
on new map areas. The only one added so far is a village, which I plan
to be occupied by townsfolk, and have a shop in it.
Added character arraylist to each mapsquare, as of now I am not sure if
I will make enemies randomly appear always, randomly appear from a set
list which gets depleted, or have fixed encounters. Certain locations
will have fixed encounters for sure.
Fixed the lake scuba gear use, and the spawn maguffin use. Temporarily
changed the loot and locations to make it easier to finsish test runs.