ZORK Update
Researched Zork Source code, which gave me some ideas about how to read
the players input.
Began work on an input parser, which takes a verb, a direct object, an
indirect object, and a preposition, which will form the basic system for
interpreting commands as actions taken on objects, with context.
Added a command class to handle these new commands.